refactor: scope component styles
