+++ components/Approach/Approach.module.scss
@@ -1,0 +1,20 @@
+.steps {
+    display: flex;
+    flex-direction: column;
+    gap: var(--space-m);
+    counter-reset: step;
+}
+
+.steps li {
+    list-style-type: none;
+    counter-increment: step;
+    padding-inline-start: var(--space-m);
+    position: relative;
+}
+
+.steps li::before {
+    content: counter(step) ".";
+    position: absolute;
+    inset-inline-start: calc(var(--space-l) * -1);
+    font-weight: 600;
+}
