+++ components/CaseExample/CaseExample.module.scss
@@ -1,0 +1,3 @@
+.caseExample {
+    margin-top: var(--space-l);
+}
