+++ components/Contact/Contact.module.scss
@@ -1,0 +1,7 @@
+.ctaGroup {
+    display: flex;
+    flex-wrap: wrap;
+    gap: var(--space-s);
+    justify-content: center;
+    margin-block-start: var(--space-2xl);
+}
