+++ components/Pledge/Pledge.module.scss
@@ -1,0 +1,28 @@
+.checklist {
+    margin: 0;
+    display: grid;
+    gap: var(--space-s);
+}
+
+.checklist div {
+    display: flex;
+    margin-top: var(--space-xs);
+    gap: var(--space-s);
+}
+
+.checklist dt {
+    font-weight: 600;
+    margin-inline-end: var(--space-xs);
+    font-family: var(--font-header), sans-serif;
+}
+
+.checklist dt::before {
+    content: "\2713";
+    color: var(--success);
+    margin-inline-end: var(--space-m);
+    margin-inline-start: var(--space-s);
+}
+
+.checklist dd {
+    margin: 0;
+}
