+++ components/Pledge/Pledge.tsx
@@ -1,5 +1,5 @@
 import Container from "@/components/Container/Container";
-import styles from "@/app/page.module.scss";
+import styles from "./Pledge.module.scss";
 
 export default function Pledge() {
     return (
