+++ components/Services/Services.module.scss
@@ -1,0 +1,11 @@
+.cards {
+    display: grid;
+    gap: var(--space-xl);
+    margin-top: var(--space-l);
+}
+
+@container (min-width:50rem) {
+    .cards {
+        grid-template-columns: repeat(2, 1fr);
+    }
+}
