+++ components/Stats/Stats.module.scss
@@ -1,0 +1,6 @@
+.stats {
+    display: flex;
+    flex-wrap: wrap;
+    gap: var(--space-xl);
+    justify-content: space-between;
+}
